+++ src/Transaction/TransactionDetailsService.php
@@ -14,6 +14,7 @@
 
 namespace HiPay\PrestaShop\Transaction;
 
+use HiPay\PrestaShop\Api\Credentials;
 use HiPay\PrestaShop\Api\PrestaShopSDK;
 use HiPay\PrestaShop\Presenter\AdminTransactionPresenter;
 
@@ -64,6 +65,19 @@ class TransactionDetailsService
                 throw new \Exception($this->module->l('Cannot retrieve transaction details', 'TransactionDetailsService'));
             }
             $order = new \Order((int) $hipayPaymentsOrder->id_order);
+            foreach ([Credentials::CREDENTIALS_TYPE_MAIN, Credentials::CREDENTIALS_TYPE_APPLE_PAY] as $credentialsType) {
+                try {
+                    $transaction = $this->sdk
+                        ->init($order->id_shop, $order->id_shop_group, $credentialsType)
+                        ->server()
+                        ->requestTransactionInformation($hipayPaymentsOrder->hipay_transaction_reference);
+                } catch (\Exception $e) {
+                    continue;
+                }
+                if (null !== $transaction) {
+                    break;
+                }
+            }
             $transaction = $this->sdk
                 ->init($order->id_shop, $order->id_shop_group)
                 ->server()